# server/docs/swagger.yaml
basePath: /
definitions:
...
  request.GetById:
    properties:
      id:
        description: 主键ID
        type: number
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      sortBy:
        type: string
      sortInverted:
        type: boolean
    type: object
...
  request.GetScheduleByTemplateId:
    properties:
      id:
        description: 主键ID
        type: number
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      sortBy:
        type: string
      sortInverted:
        type: boolean
      templateId:
        description: ProjectId  float64 `json:"projectId" form:"projectId"`
        type: number
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  scheduleMdl.Schedule:
    properties:
      createdAt:
        description: 创建时间
        type: string
      cronFormat:
        type: string
      id:
        description: 主键ID
        type: integer
      templateId:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      valid:
        description: LastCommitHash *string `gorm:"column:last_commit_hash" json:"-"`
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /task/schedule/addSchedule:
    post:
      consumes:
      - application/json
      parameters:
      - description: Schedule
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/scheduleMdl.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 新增Schedule
      tags:
      - Schedule
  /task/schedule/deleteSchedule:
    post:
      consumes:
      - application/json
      parameters:
      - description: ScheduleId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Schedule
      tags:
      - Schedule
  /task/schedule/deleteScheduleByIds:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除Schedule
      tags:
      - Schedule
  /task/schedule/getScheduleById:
    post:
      consumes:
      - application/json
      parameters:
      - description: EnvironmentId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 根据id获取Schedule
      tags:
      - Schedule
  /task/schedule/getTemplateScheduleList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.GetScheduleByTemplateId'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取基础Schedule列表
      tags:
      - Schedule
  /task/schedule/updateSchedule:
    post:
      consumes:
      - application/json
      parameters:
      - description: 主机名, 架构, 管理ip, 系统, 系统版本
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/scheduleMdl.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Schedule
      tags:
      - Schedule
  /task/schedule/validateScheduleFormat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Schedule
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/scheduleMdl.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"添加成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 确认Schedule
      tags:
      - Schedule
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"